[FIX] ge04_team2: Solve issues in PR
This modifie __manifest__.py, deleating data beacause it is not used, extend in the description.
Rename ge04_team2.py to sale_order.py and its content, the name of the class change to SaleOrder and _compute_first_purchase is refactoring to delegate the logic to _calculate_is_new_customer.
Add last line to files with no space.